rename medical-pre-authorization-ai-agent to medical-pre-authorization
