+++ python/agents/medical-pre-authorization/deployment/test_deployment.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Test deployment of Medical-Pre-Authorization-AI-Agent to Agent Engine."""
+"""Test deployment of Medical-Pre-Authorization Agent to Agent Engine."""
 
 import os
