+++ python/agents/medical-pre-authorization/tests/sample_data/patient_medical_diagnosis.pdf
